extract handler into package and create test for handler
