+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/wwgberlin/repl2go
+
+go 1.13
